ESPM-25 - organize shared components
