+++ components/SwatchesExplorer.vue
@@ -5,13 +5,13 @@ import { onMounted, onUnmounted, computed, nextTick, ref } from 'vue'
 import ColorCarousel from './carousel/ColorCarousel.vue'
 import GridControls from './grid/GridControls.vue'
 import PaletteGrid from './grid/PaletteGrid.vue'
-import PaletteControls from './shared/PaletteControls.vue'
+import PaletteControls from './palette/PaletteControls.vue'
 import AppHeader from './AppHeader.vue'
-import SavePaletteModal from './shared/SavePaletteModal.vue'
-import PaletteManagerModal from './shared/PaletteManagerModal.vue'
-import AboutModal from './shared/AboutModal.vue'
-import LoaderOverlay from './shared/LoaderOverlay.vue'
-import ToastNotification from './shared/ToastNotification.vue'
+import SavePaletteModal from './palette/SavePaletteModal.vue'
+import PaletteManagerModal from './palette/PaletteManagerModal.vue'
+import AboutModal from './general/AboutModal.vue'
+import LoaderOverlay from './ui/LoaderOverlay.vue'
+import ToastNotification from './ui/ToastNotification.vue'
 
 // Composables
 import { useColorData } from '../composables/useColorData.js'
