+++ components/palette/SavedPalettesGrid.vue
